+++ Infrastructure/Repository/Interfaces/IConstraintRepository.cs
@@ -6,11 +6,11 @@ namespace Infrastructure.Repository.Interfaces
 {
     public interface IConstraintRepository
     {
-        Task<Constraint> GetConstraintByIdAsync(int constraintId);
+        Task<Constraint> GetConstraintByIdAsync(long constraintId);
         Task<List<Constraint>> GetAllConstraintsAsync();
         Task AddConstraintAsync(Constraint constraint);
         Task UpdateConstraintAsync(Constraint constraint);
-        Task DeleteConstraintAsync(int constraintId);
+        Task DeleteConstraintAsync(long constraintId);
 
     }
 }
