+++ Infrastructure/Repository/Interfaces/IPreferenceRepository.cs
@@ -5,10 +5,10 @@ namespace Infrastructure.Repository.Interfaces
 {
     public interface IPreferenceRepository
     {
-        Task<Preference> GetPreferenceByIdAsync(int preferenceId);
+        Task<Preference> GetPreferenceByIdAsync(long preferenceId);
         Task<List<Preference>> GetAllPreferencesAsync();
         Task AddPreferenceAsync(Preference preference);
         Task UpdatePreferenceAsync(Preference preference);
-        Task DeletePreferenceAsync(int preferenceId);
+        Task DeletePreferenceAsync(long preferenceId);
     }
 }
