+++ Infrastructure/Repository/Interfaces/IRoleRepository.cs
@@ -6,10 +6,10 @@ namespace Infrastructure.Repository.Interfaces
 {
     public interface IRoleRepository
     {
-        Task<Role> GetRoleByIdAsync(int roleId);
+        Task<Role> GetRoleByIdAsync(long roleId);
         Task<List<Role>> GetAllRolesAsync();
         Task AddRoleAsync(Role role);
         Task UpdateRoleAsync(Role role);
-        Task DeleteRoleAsync(int roleId);
+        Task DeleteRoleAsync(long roleId);
     }
 }
